requirements.txt
fastapi
uvicorn
motor
pydantic
mysql-connector-python
python-multipart
uuid